app: implement fetch data from network for explore screen

+++ app/src/main/java/com/tenan/android/entity/Hotel.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class Hotel(
     @SerialName("lodging_id") val lodgingId: Int,
     @SerialName("place_name") val placeName: String,
-    @SerialName("rating") val rating: Double,
+    @SerialName("rating") val rating: Float,
     @SerialName("city") val city: String,
     @SerialName("image_url") val imageUrl: String
 )
